requirements.txt
pygame>=2.6.1
PyOpenGL>=3.1.6
numpy>=1.24.0,<2.0.0
Pillow>=9.5.0
psutil>=5.9.0
typing-extensions>=4.0.0
